@@ -8,7 +8,6 @@
   "bin": {
     "fandl": "./dist/fandl.sh"
   },
-  "type": "module",
   "engines": {
     "node": ">=18.0.0"
   },
@@ -33,25 +32,28 @@
   },
   "homepage": "https://github.com/liquid-labs/format-and-lint#readme",
   "dependencies": {
-    "@babel/eslint-parser": "^7.22.15",
+    "@babel/eslint-parser": "^7.25.1",
+    "@eslint/compat": "^1.1.1",
     "@eslint/js": "^8.56.0",
     "@liquid-labs/sdlc-resource-babel-and-rollup": "^1.0.0-alpha.9",
-    "@stylistic/eslint-plugin": "^2.6.4",
-    "eslint": "^8.50.0",
+    "@stylistic/eslint-plugin": "^2.8.0",
+    "eslint": "^9.12.0",
     "eslint-config-standard": "^17.1.0",
-    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-import": "^2.30.0",
     "eslint-plugin-jsdoc": "^48.0.4",
     "eslint-plugin-n": "^16.6.2",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^6.1.1",
+    "find-plus": "^2.0.0",
+    "gitignore-to-glob": "^0.3.0",
     "globals": "^14.0.0",
     "jest": "^29.7.0",
-    "prettier": "^3.3.3"
+    "js-yaml": "^4.1.0",
+    "prettier": "^3.3.3",
+    "standard-error-set": "^1.0.0-beta.2"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-proposal-optional-chaining": "^7.21.0",
-    "@liquid-labs/shell-toolkit": "^1.0.0-alpha.10"
+    "@liquid-labs/sdlc-resource-jest": "^1.0.0-alpha.8"
   },
   "_comply": {
     "orgKey": "@liquid-labs"
@@ -62,5 +64,8 @@
         "src/test/data/**"
       ]
     }
+  },
+  "peerDependencies": {
+    "command-line-args": "^6.0.0"
   }
 }
